% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdatetime.R
\name{as_jdatetime}
\alias{as_jdatetime}
\title{Convert an object to a jdatetime}
\usage{
as_jdatetime(x, tzone, ...)
}
\arguments{
\item{x}{a vector of jdate, POSIXct or Date.}
}
\description{
A generic function that converts other date/time classes to jdatetime
}
