% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdate.R, R/jdatetime.R
\name{is_jdate}
\alias{is_jdate}
\alias{is_jdatetime}
\title{Check an object for its class}
\usage{
is_jdate(x)

is_jdatetime(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\itemize{
\item \code{is_jdate()} checks whether an object is of class \code{jdate}.
\item \code{is_jdatetime()} checks whether an object is of class \code{jdatetime}.
}
}
\examples{
is_jdate(jdate_now() + 1) # TRUE
is_jdatetime(jdatetime_now() + as.difftime(2, units = "hours")) # TRUE
}
