% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdate.R
\name{jdate_make}
\alias{jdate_make}
\title{Construct jdate objects from individual components}
\usage{
jdate_make(year, month = 1L, day = 1L, ...)
}
\description{
Construct jdate objects from individual components
}
