% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdatetime.R
\name{jdatetime_make}
\alias{jdatetime_make}
\title{Construct jdatetime objects from from individual components}
\usage{
jdatetime_make(
  year,
  month = 1L,
  day = 1L,
  hour = 0L,
  minute = 0L,
  second = 0L,
  tzone = "",
  ...
)
}
\description{
Construct jdatetime objects from from individual components
}
