% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdate.R, R/jdatetime.R
\name{jdate_now}
\alias{jdate_now}
\alias{jdatetime_now}
\title{Current Jalali date and datetime}
\usage{
jdate_now(tzone = "")

jdatetime_now(tzone = "")
}
\arguments{
\item{tzone}{Time zone to get the current time for.}
}
\value{
\itemize{
\item \code{jdate_now()} A jdate object.
\item \code{jdatetime_now()} A jdatetime object.
}
}
\description{
System representation of the current time as jdate and jdatetime.
}
\examples{
# Current Jalali date in the local time zone
jdate_now()

# Current Jalali date in a specified time zone
jdate_now("Asia/Tokyo")

# may be true or false
jdate_now("Asia/Tehran") == jdate_now("Asia/Tokyo")

# Current Jalali datetime in the local time zone
jdatetime_now()
}
