% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty.R
\name{pretty.jdate}
\alias{pretty.jdate}
\alias{pretty.jdatetime}
\title{Pretty breakpoints for Jalali date-times}
\usage{
\method{pretty}{jdate}(x, n = 5, min.n = n\%/\%2, sep = " ", ...)

\method{pretty}{jdatetime}(x, n = 5, min.n = n\%/\%2, sep = " ", ...)
}
\arguments{
\item{x}{A vector of \code{jdate} or \code{jdatetime} objects.}

\item{n}{Integer giving the desired number of intervals.}

\item{min.n}{Non-negative integer giving the minimal number of intervals.}

\item{sep}{A character string, serving as a separator for certain formats
(e.g., between month and year).}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A vector of Jalali date-times, with two additional attributes:
\itemize{
\item "labels": A character vector of formatted labels corresponding to the breaks.
\item "format": The format specification that was used.
}
}
\description{
Generates a sequence of about n+1 equally spaced nice,
human-friendly time breaks which cover the given range.
}
\section{Implementation notes}{

This method is implemented following the logic of \code{pretty.Date()}
from the \pkg{grDevices} package in base R. It largely follows the behavior of
\code{pretty.Date()} in determining breakpoints and constructing formatted labels.
One intentional difference concerns the case \code{n = 0}. In this implementation:
\itemize{
\item If \code{length(x) = 1}, the scalar input value is returned unchanged and the format specification
is "\%Y-\%m-\%d".
\item If \code{length(x) > 1}, the function falls back to \code{pretty(x, n = 1)}.
}
}

\examples{
x <- seq(jdate("1404-01-02"), jdate("1404-03-03"), by = "week")
pretty(x, n = 5, min.n = 5)

x <- jdatetime("1404-08-21 18:06:33", "Asia/Tehran")
pretty(x, n = 5)
}
\seealso{
\code{\link[grDevices:pretty.Date]{grDevices::pretty.Date()}}
}
