% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_lables.R
\name{sh_labels}
\alias{sh_labels}
\title{Create labels for date and time components}
\usage{
sh_labels(month, month_abbr = month, weekday, weekday_abbr = weekday, am_pm)
}
\arguments{
\item{month, month_abbr}{Character vectors of length 12 giving full and abbreviated
month names, respectively, ordered from Farvardin to Esfand.}

\item{weekday, weekday_abbr}{Character vectors of length 7 giving full and
abbreviated weekday names, respectively, ordered from Saturday to Friday.}

\item{am_pm}{A character vector of length 2 giving the AM and PM labels.}
}
\value{
An object of class \code{shide_labels} that stores labels for
months, weekdays, and AM/PM.
}
\description{
\code{sh_labels()} allows users to define custom textual representations for months, weekdays
and AM/PM indicators, used by \code{shide} when parsing and formatting dates and times.
}
