% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{sh_quarter}
\alias{sh_quarter}
\title{Get the quarter of Jalali date-time objects}
\usage{
sh_quarter(x)
}
\arguments{
\item{x}{A vector of \code{jdate} or \code{jdatetime} objects.}
}
\value{
An integer vector of \code{1} to \code{4}.
}
\description{
Get the quarter of Jalali date-time objects
}
\examples{
x <- jdate("1402-12-14")
sh_quarter(x)
}
