% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{sh_tzone}
\alias{sh_tzone}
\alias{sh_tzone.jdatetime}
\title{Get the time zone component of \code{jdatetime} objects}
\usage{
sh_tzone(x)

\method{sh_tzone}{jdatetime}(x)
}
\arguments{
\item{x}{A vector of \code{jdatetime} objects.}
}
\value{
A character vector of length 1. An empty string represents the current local time zone.
}
\description{
Get the time zone component of \code{jdatetime} objects
}
\details{
This function is only valid for \code{jdatetime} objects. \code{jdate} objects do not have time zone attribute.
}
\examples{
x <- jdatetime("1402-12-14 19:13:31", tzone = "Iran")
sh_tzone(x)
}
