% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{shide-format}
\alias{shide-format}
\alias{format.jdate}
\alias{format.jdatetime}
\title{Format Jalali date-time objects}
\usage{
\method{format}{jdate}(x, format = NULL, ..., labels = NULL)

\method{format}{jdatetime}(x, format = NULL, tz = NULL, ..., labels = NULL)
}
\arguments{
\item{x}{A \code{jdate} or \code{jdatetime} vector.}

\item{format}{A character string specifying the output format. If \code{NULL},
a default format is used.}

\item{...}{These dots are for future extensions and must be empty.}

\item{labels}{An object of class \code{shide_labels} created from \code{\link[=sh_labels]{sh_labels()}}}

\item{tz}{Optional time zone for \code{jdatetime} objects. If provided, the
date-time is formatted as if displayed in this time zone.}
}
\value{
A character vector representing the formatted dates or date-times.
}
\description{
Format objects of class \code{jdate} and \code{jdatetime} into character vectors
using a specified format string and optional localized labels.
}
\seealso{
\code{\link[base:strptime]{base::strftime()}}
}
