% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shide-package.R
\docType{package}
\name{shide-package}
\alias{shide}
\alias{shide-package}
\title{shide: Date/Time Classes Based on Jalali Calendar}
\description{
Implements S3 classes for storing dates and date-times based on Jalali calendar.
}
\author{
\strong{Maintainer}: Mohsen Mollayi \email{mmollayi@gmail.com}

}
\keyword{internal}
