% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdate.R, R/jdatetime.R, R/shide-package.R
\name{vec_ptype2.jdate}
\alias{vec_ptype2.jdate}
\alias{vec_cast.jdate}
\alias{vec_arith.jdate}
\alias{vec_math.jdate}
\alias{vec_ptype2.jdatetime}
\alias{vec_cast.jdatetime}
\alias{vec_arith.jdatetime}
\alias{vec_math.jdatetime}
\alias{shide-vctrs}
\title{Internal vctrs methods}
\usage{
\method{vec_ptype2}{jdate}(x, y, ...)

\method{vec_cast}{jdate}(x, to, ...)

\method{vec_arith}{jdate}(op, x, y, ...)

\method{vec_math}{jdate}(.fn, .x, ...)

\method{vec_ptype2}{jdatetime}(x, y, ...)

\method{vec_cast}{jdatetime}(x, to, ...)

\method{vec_arith}{jdatetime}(op, x, y, ...)

\method{vec_math}{jdatetime}(.fn, .x, ...)
}
\description{
These methods are the extensions that allow shide objects to
work with vctrs.
}
\keyword{internal}
