% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdate.R
\name{vec_cast.jdate}
\alias{vec_cast.jdate}
\title{Cast an object to a \code{jdate} object}
\usage{
\method{vec_cast}{jdate}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}
}
\value{
A vector of \code{jdate} objects.
}
\description{
Cast an object to a \code{jdate} object
}
\seealso{
\link{as_jdate} is a convenience function that makes use of the casts that
are defined for \code{vec_cast.jdate()} methods.
}
