% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdate.R
\name{is_jdate}
\alias{is_jdate}
\title{is_jdate()
`is_jdate()` checks whether an object is of class `jdate`.}
\usage{
is_jdate(x)
}
\description{
is_jdate()
`is_jdate()` checks whether an object is of class `jdate`.
}
