% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{seq.jdate}
\alias{seq.jdate}
\title{Sequence of Jalali dates}
\usage{
\method{seq}{jdate}(from, to, by, length.out = NULL, along.with = NULL, ...)
}
\arguments{
\item{from}{starting date.  Required}

\item{to}{end date.  Optional.}

\item{by}{increment of the sequence.  Optional.  See \sQuote{Details}.}

\item{length.out}{integer, optional.  Desired length of the sequence.}

\item{along.with}{take the length from the length of this argument.}

\item{...}{arguments passed to or from other methods.}
}
\value{
A vector of \code{jdate} objects.
}
\description{
The method for \link{seq} for objects of class \code{jdate}. This method is a wrapper around \code{base::seq.Date()}
}
\details{
\code{by} can be specified in several ways:
\itemize{
\item A number, taken to be in days.
\item A object of class \link{difftime}.
\item A character string, containing one of "day", "week", "month", "quarter" or "year".
This can optionally be preceded by a (positive or negative) integer
and a space, or followed by "s".
}
}
\examples{
## by days
seq(jdate("1402-01-01"), jdate("1402-01-10"), 1)
## by 2 weeks
seq(jdate("1402-01-01"), jdate("1402-04-01"), "2 weeks")
## first days of years
seq(jdate("1390-01-01"), jdate("1399-01-01"), "years")
## by month
seq(jdate("1400-01-01"), by = "month", length.out = 12)
## quarters
seq(jdate("1400-01-01"), jdate("1403-01-01"), by = "quarter")
}
