% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shide-package.R
\docType{package}
\name{shide-package}
\alias{shide}
\alias{shide-package}
\title{shide: Date/Time Classes Based on Jalali Calendar}
\description{
Implements S3 classes for storing dates and date-times based on the Jalali calendar. The main design goal of 'shide' is consistency with base R's 'Date' and 'POSIXct'. It provide features such as: date-time parsing, formatting and arithmetic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mmollayi/shide}
  \item Report bugs at \url{https://github.com/mmollayi/shide/issues}
}

}
\author{
\strong{Maintainer}: Mohsen Mollayi \email{mmollayi@gmail.com} [copyright holder]

}
\keyword{internal}
