% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shide-package.R
\docType{package}
\name{shide-package}
\alias{shide}
\alias{shide-package}
\title{shide: Date/Time Classes Based on Jalali Calendar}
\description{
Implements S3 classes for storing dates and date-times based on Jalali calendar.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mmollayi/shide}
  \item Report bugs at \url{https://github.com/mmollayi/shide/issues}
}

}
\author{
\strong{Maintainer}: Mohsen Mollayi \email{mmollayi@gmail.com} [copyright holder]

}
\keyword{internal}
